refactor: export transformers from index - #397
OOXML spacing changes
