+++ packages/markdown-docx/src/CiceroMarkToOOXML/rules.js
@@ -14,7 +14,7 @@
 
 'use strict';
 
-const { sanitizeHtmlChars } = require('./CiceroMarkToOOXMLHelpers');
+const { sanitizeHtmlChars } = require('./helpers');
 
 /**
  * Inserts text.
@@ -23,7 +23,9 @@ const { sanitizeHtmlChars } = require('./CiceroMarkToOOXMLHelpers');
  * @returns {string} OOXML tag for the text
  */
 const TEXT_RULE = (value) => {
-    return `<w:r><w:t xml:space="preserve">${sanitizeHtmlChars(value)}</w:t></w:r>`;
+    return `<w:r>
+        <w:t xml:space="preserve">${sanitizeHtmlChars(value)}</w:t>
+    </w:r>`;
 };
 
 /**
@@ -33,7 +35,12 @@ const TEXT_RULE = (value) => {
  * @returns {string} OOXML tag for the emphasised text
  */
 const EMPHASIS_RULE = (value) => {
-    return `<w:r><w:rPr><w:i /></w:rPr><w:t>${sanitizeHtmlChars(value)}</w:t></w:r>`;
+    return `<w:r>
+        <w:rPr>
+            <w:i />
+        </w:rPr>
+        <w:t>${sanitizeHtmlChars(value)}</w:t>
+    </w:r>`;
 };
 
 module.exports = { TEXT_RULE, EMPHASIS_RULE };
